Fixed problems with uploaded files
File Uploader was being triggered every time a widget was changed (Streamlit-side bug), which was resetting lots of variables and preventing user selection of plot variables. This is only apparent when using the File Uploader button, so I'd missed it by automatically loading the Sample Data Set. Code is still messy, but will be cleaned up in next version.